% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqsurface-package.R
\docType{package}
\name{aqsurface-package}
\alias{aqsurface}
\alias{aqsurface-package}
\title{aqsurface: spatial interpolation surfaces for urban air quality}
\description{
Reproducible workflow for comparing universal kriging (\code{\link[gstat:krige]{gstat::krige()}})
with generalised additive models (\code{\link[mgcv:gam]{mgcv::gam()}}) on hourly urban air
pollutant concentrations.
}
\details{
Typical pipeline:
\enumerate{
\item \code{\link[=read_pollutant_rdata]{read_pollutant_rdata()}} / \code{\link[=load_pollutant]{load_pollutant()}} to ingest the
legacy \code{pm10.RData} / \code{no2.RData} archives.
\item \code{\link[=make_station_sf]{make_station_sf()}} / \code{\link[=split_stations]{split_stations()}} to align stations.
\item \code{\link[=fit_uk]{fit_uk()}} or \code{\link[=fit_gam]{fit_gam()}} for one column; \code{\link[=predict_surface_batch]{predict_surface_batch()}}
to loop over a decade of day/night columns.
\item \code{\link[=apply_road_ratio]{apply_road_ratio()}} (optional) for road-side correction.
\item \code{\link[=extract_at_stations]{extract_at_stations()}} + \code{\link[=compute_metrics]{compute_metrics()}} for validation
against the held-out road monitoring stations.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dataandcrowd/aqsurface}
  \item Report bugs at \url{https://github.com/dataandcrowd/aqsurface/issues}
}

}
\author{
\strong{Maintainer}: Hyesop Shin \email{hyesop.shin@auckland.ac.nz}

}
\keyword{internal}
