% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{predict_gam}
\alias{predict_gam}
\title{Predict a GAM surface onto a regular grid}
\usage{
predict_gam(object, newgrid, se_fit = FALSE)
}
\arguments{
\item{object}{Result from \code{\link[=fit_gam]{fit_gam()}}.}

\item{newgrid}{An \code{sf}, \link[terra:SpatRaster-class]{terra::SpatRaster} or \code{data.frame} with
\code{X}, \code{Y} columns. Rasters are converted to cell centres.}

\item{se_fit}{If \code{TRUE}, also return prediction standard errors.}
}
\value{
A list with \code{pred} (and optionally \code{se}) and, if \code{newgrid} was
a raster, \code{raster} and \code{raster_se}.
}
\description{
Predict a GAM surface onto a regular grid
}
